moved and updated script, denylist nit picks
