+++ packages/syft-libs/pyscaffoldext-syft-support/setup.cfg
@@ -48,6 +48,7 @@ package_dir =
 install_requires =
     importlib-metadata; python_version<"3.8"
     pyscaffold>=4.0,<5.0a0
+    typing-inspect
 
 
 [options.packages.find]
