+++ packages/syft/src/syft/lib/__init__.py
@@ -265,7 +265,6 @@ lib_ast = create_lib_ast(None)
 @wrapt.when_imported("statsmodels")
 @wrapt.when_imported("sympc")
 @wrapt.when_imported("tenseal")
-#@wrapt.when_imported("xgboost")
 @wrapt.when_imported("zksk")
 @wrapt.when_imported("pytorch_lightning")
 @wrapt.when_imported("transformers")
